+++ src/main/java/org/folio/edge/patron/Constants.java
@@ -4,15 +4,24 @@ import java.util.Currency;
 
 public class Constants {
 
+  public static final String KEYCLOAK_URL = "KC_URL";
   public static final String SYS_PATRON_ID_CACHE_TTL_MS = "patron_id_cache_ttl_ms";
   public static final String SYS_NULL_PATRON_ID_CACHE_TTL_MS = "null_patron_id_cache_ttl_ms";
   public static final String SYS_PATRON_ID_CACHE_CAPACITY = "patron_id_cache_capacity";
+  public static final String SYS_KEYCLOAK_KEY_CACHE_TTL_MS = "keycloak_key_cache_ttl_ms";
+  public static final String SYS_NULL_KEYCLOAK_KEY_CACHE_TTL_MS = "null_keycloak_key_cache_ttl_ms";
+  public static final String SYS_KEYCLOAK_KEY_CACHE_CAPACITY = "keycloak_key_cache_capacity";
   public static final String PARAM_EXPIRED = "expired";
+  public static final String VIP_CLAIM = "vip";
+  public static final String EXTERNAL_SYSTEM_ID_CLAIM = "externalSystemId";
 
   public static final String DEFAULT_CURRENCY_CODE = Currency.getInstance("USD").getCurrencyCode();
   public static final long DEFAULT_PATRON_ID_CACHE_TTL_MS = 60 * 60 * 1000L;
   public static final long DEFAULT_NULL_PATRON_ID_CACHE_TTL_MS = 30 * 1000L;
   public static final int DEFAULT_PATRON_ID_CACHE_CAPACITY = 1000;
+  public static final long DEFAULT_KEYCLOAK_KEY_CACHE_TTL_MS = 60 * 60 * 1000L;
+  public static final long DEFAULT_NULL_KEYCLOAK_KEY_CACHE_TTL_MS = 30 * 1000L;
+  public static final int DEFAULT_KEYCLOAK_KEY_CACHE_CAPACITY = 50;
 
   public static final String PARAM_SORT_BY = "sortBy";
   public static final String PARAM_LIMIT = "limit";
@@ -25,6 +34,7 @@ public class Constants {
   public static final String PARAM_INSTANCE_ID = "instanceId";
   public static final String PARAM_HOLD_ID = "holdId";
   public static final String PARAM_EMAIL_ID = "emailId";
+  public static final String PARAM_EXTERNAL_SYSTEM_ID = "externalSystemId";
   public static final String PARAM_REQUEST_ID = "requestId";
 
   public static final String MSG_ACCESS_DENIED = "Access Denied";